refactor: часть логики работы с жанрами вынесена в отдельный сервис.
