+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.service.GenreService;
 
 import java.util.List;
 
@@ -16,17 +17,17 @@ import java.util.List;
 @RequestMapping("/genres")
 @RequiredArgsConstructor
 public class GenreController {
-    private final FilmService filmService;
+    private final GenreService genreService;
 
     @GetMapping
     public List<Genre> getAll() {
         log.info("GET /genres");
-        return filmService.getAllGenres();
+        return genreService.getAllGenres();
     }
 
     @GetMapping("/{id}")
     public Genre getGenreById(@PathVariable(name = "id") int id) {
         log.info(String.format("GET /genres/{id}, {id} = %s", id));
-        return filmService.getGenreById(id);
+        return genreService.getGenreById(id);
     }
 }
