+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -1,0 +1,31 @@
+package ru.yandex.practicum.filmorate.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.storage.GenreStorage;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class GenreService {
+
+    protected final GenreStorage genreStorage;
+
+    public List<Genre> getAllGenres() {
+        return genreStorage.getAllGenres();
+    }
+
+    public Genre getGenreById(int id) {
+        final Optional<Genre> genreOpt = genreStorage.getGenreById(id);
+
+        if (genreOpt.isEmpty()) {
+            throw new NotFoundException(String.format("Жанр с id = %s не найден", id));
+        }
+
+        return genreOpt.get();
+    }
+}
